Moving the changes into a single file, keeping routers.py tidy
